+++ packages/gatsby-recipes/src/renderer/render.js
@@ -1,4 +1,4 @@
-import React, { Suspense, useContext, useState } from "react"
+import React, { Suspense, useContext } from "react"
 import Queue from "better-queue"
 import lodash from "lodash"
 import mitt from "mitt"
@@ -15,6 +15,7 @@ import {
 import { useRecipeStep } from "./step-component"
 import { InputProvider } from "./input-provider"
 import { ResourceProvider, useResourceContext } from "./resource-provider"
+import findDependencyMatch from "../find-dependency-match"
 
 const errorCache = new Map()
 
@@ -28,45 +29,34 @@ const getUserProps = props => {
   return userProps
 }
 
-const SetResourcesProvider = React.createContext()
-
-let resourcesCache
-
 const Wrapper = ({
   children,
   inputs,
   isApply,
   resultCache,
   inFlightCache,
+  blockedResources,
   queue,
-}) => {
-  // eslint-disable-next-line
-  const [resourcesList, setResources] = useState(resourcesCache || [])
-  resourcesCache = resourcesList
-
-  return (
-    <ErrorBoundary>
-      <GlobalsProvider
-        value={{
-          mode: isApply ? `apply` : `plan`,
-          resultCache,
-          inFlightCache,
-          queue,
-        }}
-      >
-        <SetResourcesProvider.Provider value={setResources}>
-          <ResourceProvider value={resourcesList}>
-            <InputProvider value={inputs}>
-              <Suspense fallback={<p>Loading recipe...</p>}>
-                {children}
-              </Suspense>
-            </InputProvider>
-          </ResourceProvider>
-        </SetResourcesProvider.Provider>
-      </GlobalsProvider>
-    </ErrorBoundary>
-  )
-}
+  plan,
+}) => (
+  <ErrorBoundary>
+    <GlobalsProvider
+      value={{
+        mode: isApply ? `apply` : `plan`,
+        resultCache,
+        inFlightCache,
+        blockedResources,
+        queue,
+      }}
+    >
+      <ResourceProvider value={plan}>
+        <InputProvider value={inputs}>
+          <Suspense fallback={<p>Loading recipe...</p>}>{children}</Suspense>
+        </InputProvider>
+      </ResourceProvider>
+    </GlobalsProvider>
+  </ErrorBoundary>
+)
 
 const ResourceComponent = ({
   _resourceName: Resource,
@@ -75,7 +65,13 @@ const ResourceComponent = ({
   children,
   ...props
 }) => {
-  const { mode, resultCache, inFlightCache, queue } = useGlobals()
+  const {
+    mode,
+    resultCache,
+    inFlightCache,
+    blockedResources,
+    queue,
+  } = useGlobals()
   const step = useRecipeStep()
   const parentResourceContext = useParentResourceContext()
 
@@ -89,6 +85,7 @@ const ResourceComponent = ({
       mode,
       resultCache,
       inFlightCache,
+      blockedResources,
       queue,
     },
     props
@@ -121,7 +118,7 @@ const validateResource = (resourceName, context, props) => {
 
 const handleResource = (resourceName, context, props) => {
   // Initialize
-  const { mode, resultCache, inFlightCache, queue } = context
+  const { mode, resultCache, inFlightCache, blockedResources, queue } = context
 
   // TODO use session ID to ensure the IDs are unique..
   const trueKey = props._key ? props._key : context._uuid
@@ -134,26 +131,11 @@ const handleResource = (resourceName, context, props) => {
     cacheKey = JSON.stringify({ resourceName, ...props, mode })
   }
 
-  // update global context when results come in.
-  const updateResource = result => {
-    allResources = allResources.filter(a => a.resourceDefinitions._key)
-    const resourceMap = new Map()
-
-    allResources.forEach(r => resourceMap.set(r.resourceDefinitions._key, r))
-    const newResource = {
-      resourceName,
-      resourceDefinitions: props,
-      ...result,
-    }
-
-    if (!lodash.isEqual(newResource, resourceMap.get(trueKey))) {
-      resourceMap.set(trueKey, newResource)
-      // TODO Do we need this? It's causing infinite loops
-      // setResources([...resourceMap.values()])
-    }
-  }
-
   let allResources = useResourceContext()
+  const resourcePlan = allResources?.find(
+    a => a.resourceDefinitions._key === trueKey || a._uuid === trueKey
+  )
+
   if (!errorCache.has(trueKey)) {
     const error = validateResource(resourceName, context, props)
     errorCache.set(trueKey, error)
@@ -161,7 +143,6 @@ const handleResource = (resourceName, context, props) => {
       const result = {
         error: `Validation error: ${error.details[0].message}`,
       }
-      updateResource(result)
       resultCache.set(cacheKey, result)
       return result
     }
@@ -171,12 +152,41 @@ const handleResource = (resourceName, context, props) => {
   const inFlightPromise = inFlightCache.get(cacheKey)
 
   if (cachedResult) {
-    updateResource(cachedResult)
     return cachedResult
   }
 
   if (inFlightPromise) {
     throw inFlightPromise
+  }
+
+  // If this resource requires another resource to be created before it,
+  // check to see if they're created. The first time this is called,
+  // create a promise which we cache and keep throwing until
+  // all the dependencies are created & then we reject the promise
+  // to trigger andother render where we finally can create the
+  // now no-longer-blocked resource.
+  //
+  // TODO test this when we can mock resources by varying what
+  // resources depend on what & which return first and ensuring
+  // resources end in right order.
+  if (mode === `apply` && resourcePlan.dependsOn) {
+    const matches = findDependencyMatch(allResources, resourcePlan)
+    let outsideReject
+    if (!matches.every(m => m.isDone)) {
+      // Probably we're going to need a state machine
+      // just for installing things, sheesh.
+      if (blockedResources.get(cacheKey)) {
+        throw blockedResources.get(cacheKey).promise
+      }
+      const promise = new Promise((resolve, reject) => {
+        outsideReject = reject
+      })
+      blockedResources.set(cacheKey, { promise, outsideReject })
+      throw promise
+    } else {
+      blockedResources.get(cacheKey).outsideReject()
+      blockedResources.delete(cacheKey)
+    }
   }
 
   const fn = mode === `apply` ? `create` : `plan`
@@ -190,14 +200,12 @@ const handleResource = (resourceName, context, props) => {
       const cachedValue = resultCache.get(cacheKey)
       if (cachedValue) {
         resolve(cachedValue)
-        updateResource(cachedValue)
       } else {
         resources[resourceName][fn](context, props)
           .then(result => {
             if (fn === `create`) {
               result.isDone = true
             }
-            updateResource(result)
             inFlightCache.set(cacheKey, false)
             return result
           })
@@ -225,9 +233,10 @@ const handleResource = (resourceName, context, props) => {
   throw promise
 }
 
-const render = (recipe, cb, inputs = {}, isApply, isStream, name) => {
+const render = (recipe, cb, context = {}, isApply, isStream, name) => {
+  const { inputs } = context
   const emitter = mitt()
-  const plan = {}
+  const renderState = {}
 
   const queue = new Queue(
     async (job, cb) => {
@@ -239,15 +248,19 @@ const render = (recipe, cb, inputs = {}, isApply, isStream, name) => {
 
   const resultCache = new Map()
   const inFlightCache = new Map()
+  const blockedResources = new Map()
 
   let result
+  let resourcesArray = []
 
   const recipeWithWrapper = (
     <Wrapper
       inputs={inputs}
+      plan={context.plan}
       isApply={isApply}
       resultCache={resultCache}
       inFlightCache={inFlightCache}
+      blockedResources={blockedResources}
       queue={queue}
     >
       {recipe}
@@ -257,9 +270,9 @@ const render = (recipe, cb, inputs = {}, isApply, isStream, name) => {
   // Keep calling render until there's remaining resources to render.
   // This let's resources that depend on other resources to pause until one finishes.
   const renderResources = isDrained => {
-    result = RecipesReconciler.render(recipeWithWrapper, plan, name)
+    result = RecipesReconciler.render(recipeWithWrapper, renderState, name)
 
-    const resourcesArray = transformToPlan(result)
+    resourcesArray = transformToPlan(result)
 
     // Tell UI about updates.
     emitter.emit(`update`, resourcesArray)
@@ -279,7 +292,11 @@ const render = (recipe, cb, inputs = {}, isApply, isStream, name) => {
       ) {
         result = true
         // If there's still nothing on the queue and we've drained the queue, that means we're done.
-      } else if (isDrained && queue.length === 0) {
+      } else if (
+        isDrained &&
+        queue.length === 0 &&
+        blockedResources.size === 0
+      ) {
         result = true
         // If there's one resource & it fails validation, it doesn't go into the queue
         // so we check if inFlightCache is empty & all resources have an error.